pyproject.toml
[project]
name = "spxtacular"
version = "0.2.0"
description = "Mass spectrometry spectrum processing library."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
keywords = ["mass spectrometry", "proteomics", "deconvolution", "bioinformatics", "peptide"]
dependencies = [
    "numpy",
    "pandas>=2.0",
    "peptacular>=3.1.0",
    "paftacular[sequence]>=1.0.0",
    "tdfpy>=1.0.2",
    "plotly>=6.6.0",
    "mzmlpy>=0.3.1",
]
authors = [
    { name = "Patrick Garrett", email = "pgarrett@scripps.edu" }
]
maintainers = [
    { name = "Patrick Garrett", email = "pgarrett@scripps.edu" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]

[project.urls]
Homepage = "https://github.com/tacular-omics/spxtacular"
Repository = "https://github.com/tacular-omics/spxtacular"
Documentation = "https://tacular-omics.github.io/spxtacular"
"Bug Tracker" = "https://github.com/tacular-omics/spxtacular/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spxtacular"]

[project.optional-dependencies]
numba = ["numba>=0.64.0"]

[tool.uv]
package = true
dev-dependencies = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.5",
    "numba>=0.64.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.11",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

#[tool.uv.sources]
#peptacular = { path = "../peptacular", editable = true }
#tdfpy = { path = "../tdfpy", editable = false }
#mzmlpy = { path = "../mzmlpy", editable = true }
